chore: move aws and datadog out of contrib crate.
Added opentelemetry-jaeger, opentelemtry-zipkin, opentelemetry-datadog, opentelemetry-aws into code coverage check.
