+++ .github/workflows/ci.yml
@@ -71,7 +71,7 @@ jobs:
     - uses: actions-rs/cargo@v1
       with:
         command: test
-        args: -p opentelemetry --all-features --no-fail-fast
+        args: -p opentelemetry -p opentelemetry-jaeger -p opentelemtry-zipkin -p opentelemetry-datadog -p opentelemetry-aws --all-features --no-fail-fast
       env:
         CARGO_INCREMENTAL: '0'
         RUSTFLAGS: '-Zprofile -Ccodegen-units=1 -Cinline-threshold=0 -Clink-dead-code -Coverflow-checks=off -Cpanic=abort -Zpanic_abort_tests'
