+++ examples/datadog/Cargo.toml
@@ -5,4 +5,4 @@ edition = "2018"
 
 [dependencies]
 opentelemetry = { path = "../../opentelemetry" }
-opentelemetry-contrib = { path = "../../opentelemetry-contrib", features = ["datadog", "reqwest-blocking-client"] }
+opentelemetry-datadog = { path = "../../opentelemetry-datadog", features = ["reqwest-blocking-client"] }
