+++ opentelemetry-aws/CHANGELOG.md
@@ -1,0 +1,5 @@
+## v0.1.0
+
+### Added
+
+- AWS XRay propagator
