+++ opentelemetry-aws/src/lib.rs
@@ -1,0 +1,8 @@
+//! # Opentelemetry AWS Integration
+//!
+//! This crate provides integration with different telemetry platform.
+//!
+//!
+
+#[cfg(feature = "trace")]
+pub mod trace;
