+++ opentelemetry-aws/src/trace/mod.rs
@@ -1,0 +1,3 @@
+mod xray;
+
+pub use xray::propagator::XrayPropagator;
