+++ opentelemetry-aws/src/trace/xray/mod.rs
@@ -1,0 +1,1 @@
+pub mod propagator;
