+++ opentelemetry-contrib/src/trace/mod.rs
@@ -1,5 +1,4 @@
 //! # Opentelemetry trace contrib
 //!
 
-pub mod exporter;
 pub mod propagator;
