+++ opentelemetry-datadog/CHANGELOG.md
@@ -1,0 +1,6 @@
+## v0.1.0
+
+### Added
+
+- Datadog exporter
+- Datadog propagator
